Manager.Args

Description

|

Documentation

data Options

All data collated from command line arguments

Constructors

Options

Fields

workDir :: FilePath

Work directory argument Where the VHDL WORK library files are stored Arg: "--work-dir" Default: "work/"

ieeeDir :: FilePath

IEEE directory argument Where the VHDL IEEE library files are stored Arg: "--ieee-dir" Default: "ieee/"

buildDir :: FilePath

Build directory argument Where the tool outputs simulation files to Arg: "--build-dir" Default: "build/"

topModule :: String

Top module argument The name of the top level VHDL design entity Arg: "--top" Default: N/A

optionsGroup :: Parser Options

Command line argument parser For details of the argument data stored, see type Options