Parser.Combinators.Lex

Description

|

Documentation

identifier :: Parser UpperString

Parses an identifier Returns upper case string form of the identifier (UpperString) identifier ::= letter { [ underliner ] letter_or_digit }

abstractLiteral :: Parser AbstractLiteral

Parses an abstract literal Returns the parsed AbstractLiteral value abstract_literal ::= decimal_literal | based_literal decimal_literal ::= integer [ . integer ] [ exponent ] based_literal ::= base [ exponent ] base ::= integer based_integer ::= extended_digit { [ underline ] extended_digit } extended_digit ::= digit | letter NOTE: - Underline is ignored - Base determines valid extended_digit set - extended_digit is case insensitive

characterLiteral :: Parser Char

Parses a character literal Returns the Char literal (without the containing ''' characters) character_literal ::= ' graphic_character '

stringLiteral :: Parser String

Parses a string literal Returns the String literal (without the containing '"' characters) string_literal ::= " { graphic_character } " NOTE: - Can use character % instead of '"' - To use the container character within the string, must repeat the character

bitStringLiteral :: Parser BitString

Parses a bit string literal Returns the parsed bit string literal bit_string_literal ::= base_specifier " bit_value " bit_value ::= extended_digit { [ underline ] extended_digit } base_specifier ::= B | O | X NOTE: - Base specifier is case insensitive - Can use character % instead of '"' - Extended digit depends on base specifier

If base specifier == B: extended_digit ::= 0 | 1 If base specifier == O: extended_digit ::= 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 If base specifier == X: extended_digit ::= 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | A | B | C | D | E | F NOTE: - Extended digit is case insensitive

comment :: Parser ()

Parses a comment Returns '()' - Comments start with the string "--" - Comments continue to the end of a line