Parser.Combinators.Lex.Internal

Description

|

Documentation

graphicCharacter :: Parser Char

Parse a graphic character Returns parsed character graphic_character ::= basic_graphic_character | lower_case_letter | other_special_character

basicGraphicCharacter :: Parser Char

Parse a basic graphic character Returns parsed character basic_graphic_character ::= upper_case_letter | digit | special_character | space_character

basicCharacter :: Parser Char

Parse a basic character Returns parsed character basic_character ::= basic_graphic_character | format_effector

upperCaseLetter :: Parser Char

Parse an upper case letter Returns parsed character upper case letters A B C D E F G H I J K L M N O P Q R S T U V W X Y Z

specialCharacter :: Parser Char

Parse a special character Returns parsed character special characters " # & ' ( ) * + , - . / : ; = _ |

spaceCharacter :: Parser Char

Parse the space character Returns the space character

formatEffector :: Parser Char

Parse a format effector Returns parsed character

lowerCaseLetter :: Parser Char

Parse a lower case letter Returns parsed character upper case letters a b c d e f g h i j k l m n o p q r s t u v w x y z

otherSpecialCharacter :: Parser Char

Parse an other special character Returns parsed character other special characters ! $ % ? [ ] ^ ` { } ~ @

letterOrDigit :: Parser Char

Parses a letter or digit Returns parsed character letter_or_digit ::= letter | digit

letter :: Parser Char

Parses a letter Returns parsed character letter ::= upper_case_letter | lower_case_letter

integer :: Parser String

Parses an integer Returns the String of the integer integer ::= digit { [ underline ] digit }

exponent' :: Parser Integer

Parses an exponent Returns the Integer exponent value exponent ::= E [ + ] integer | E - integer NOTE: - Case insensitive E