Parser.Combinators.ReservedWords.Internal

Description

|

Documentation

caseInsensitiveChar :: Char -> Parser Char

Parse a character, ignoring case Parses Char input (upper or lower case) Returns upper case character parsed

keywordParser :: String -> Parser ()

Parses a keyword, ignoring case Parses String input (upper or lower case) Returns '()'