Parser.Types.Token

Description

|

Documentation

data UpperString

Upper case string data Only permitted to contain upper case characters

mkUpperString :: String -> UpperString

Upper case string constructor Converts string to upper case and packs it into a bytestring

data AbstractLiteral

Abstract literal value abstract_literal ::= decimal_literal | based_literal = NOTE Lexed output does not match tree specification There are two acceptable types of this literal * Decimal Literals * Based Literals The lexer converts both of these to their representative value: * Integer value * Real - Floating point value

Constructors

UniversalInteger Int64
Universal integer type value Implemented with Haskell Int64 type

UniversalReal Double
Universal real type value Implemented with Haskell Double type

data BitString

Bit string data Only permitted to contain characters '0' and '1'

mkBitString :: String -> BitString

Bit string constructor Forces string to contain '0' and '1' and packs it into a bytestring Error if string contains any characters other than '0' or '1'