Parser.Types.Token.Internal

Description

|

Documentation

newtype UpperString

Upper case string data Only permitted to contain upper case characters

Constructors

UpperString ByteString

newtype BitString

Bit string data Only permitted to contain characters '0' and '1'

Constructors

BitString ByteString